test: support optional plugin download on test:setup

+++ bin/install-wp-tests.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 
 if [ $# -lt 3 ]; then
-	echo "usage: $0 <db-name> <db-user> <db-pass> [db-host] [wp-version] [skip-database-creation]"
+	echo "usage: $0 <db-name> <db-user> <db-pass> [db-host] [wp-version] [skip-plugin-download] [skip-database-creation]"
 	exit 1
 fi
 
@@ -10,7 +10,8 @@ DB_USER=$2
 DB_PASS=$3
 DB_HOST=${4-localhost}
 WP_VERSION=${5-latest}
-SKIP_DB_CREATE=${6-false}
+SKIP_PLUGIN_DOWNLOAD=${6-false}
+SKIP_DB_CREATE=${7-false}
 
 TMPDIR=${TMPDIR-/tmp}
 TMPDIR=$(echo $TMPDIR | sed -e "s/\/$//")
@@ -177,18 +178,13 @@ install_db() {
 }
 
 install_acf() {
-	REPO_NAME="wp-paid-plugins"
-
-	rm -rf $TMPDIR/$REPO_NAME
-	rm -rf "${LOCALTMPDIR}/advanced-custom-fields-pro"
-
-	if [[ -z "${GITHUB_TOKEN}" ]]; then
-		git clone https://$GITHUB_TOKEN@github.com/helsingborg-stad/$REPO_NAME.git $TMPDIR/$REPO_NAME
-	else
+	if [[ $SKIP_PLUGIN_DOWNLOAD = "false" ]]; then
+		REPO_NAME="wp-paid-plugins"
+		rm -rf $TMPDIR/$REPO_NAME
+		rm -rf "${LOCALTMPDIR}/advanced-custom-fields-pro"
 		git clone git@github.com:helsingborg-stad/$REPO_NAME.git $TMPDIR/$REPO_NAME
+		unzip $TMPDIR/$REPO_NAME/acf.zip -d $TMPDIR
 	fi
-	
-	unzip $TMPDIR/$REPO_NAME/acf.zip -d $TMPDIR
 }
 
 install_acf
